Enhanced transformation effiency
* Inplace transformation convention
* @transformation decorator to apply the transformation on a new copy of the GraphModule
* Added a "lint_and_recompile" flag to inplace transformations to avoid multiple lint and recompile when applying multiple transformations as it is very slow otherwise
* Used all of those additions to make "transform_to_dynamic_input_" much faster
